Check the .msi path

+++ generateMsiFiles.ps1
@@ -19,7 +19,8 @@ if (-not (@($env:Path -split ";") -contains $env:WIX))
 # Get runtime version
 $artifactsPath = "$baseDir\artifacts"
 $buildDir = "$baseDir\build"
-$cli = Get-ChildItem -Path $artifactsPath -Include func.dll -Recurse | Select-Object -First 1
+$cli = (Get-ChildItem -Path $artifactsPath -Include func.dll -Recurse | Select-Object -First 1).FullName
+Write-Host $cli
 $cliVersion = [System.Diagnostics.FileVersionInfo]::GetVersionInfo($cli).FileVersion
 
 # Generate MSI installers for Windows
@@ -41,12 +42,15 @@ $cliVersion = [System.Diagnostics.FileVersionInfo]::GetVersionInfo($cli).FileVer
     $fragmentPath = "$buildDir\$fragmentName.wxs"
     $msiPath = "$artifactsPath\$msiName.msi"
 
-    & {
-        heat dir '.' -cg FuncHost -dr INSTALLDIR -gg -ke -out $fragmentPath -srd -sreg -template fragment -var var.Source
-        candle -arch $platform -dPlatform="$platform" -dSource='.' -dProductVersion="$cliVersion" $masterWxsPath $fragmentPath
-        light -ext WixUIExtension -out $msiPath -sice:ICE61 $masterWxsName.wixobj $fragmentName.wixobj
+    & { heat dir '.' -cg FuncHost -dr INSTALLDIR -gg -ke -out $fragmentPath -srd -sreg -template fragment -var var.Source }
+    & { candle -arch $platform -dPlatform="$platform" -dSource='.' -dProductVersion="$cliVersion" $masterWxsPath $fragmentPath }
+    & { light -ext "WixUIExtension" -out $msiPath -sice:"ICE61" "$masterWxsName.wixobj" "$fragmentName.wixobj" }
+    
+    # Check that the .msi files are actually present
+    if (-not(Test-Path -Path $msiPath))
+    {
+        throw "$msiPath not found."
     }
-    if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)}
 
     Set-Location $baseDir
     Get-ChildItem -Path $targetDir -Recurse | Remove-Item -Force -Recurse -ea SilentlyContinue
